Moving proper error into networking module
