+++ features/facts/build.gradle.kts
@@ -10,6 +10,7 @@ dependencies {
     implementation(project(":platform:shared-assets"))
     implementation(project(":platform:shared-utilities"))
     implementation(project(":platform:navigator"))
+    implementation(project(":platform:networking"))
 
     implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.5.20")
     implementation("org.kodein.di:kodein-di-jvm:7.6.0")
