+++ features/facts/src/main/java/io/dotanuki/norris/facts/ErrorStateResources.kt
@@ -1,8 +1,8 @@
 package io.dotanuki.norris.facts
 
-import io.dotanuki.norris.domain.errors.NetworkingError
-import io.dotanuki.norris.domain.errors.RemoteServiceIntegrationError.RemoteSystem
 import io.dotanuki.norris.domain.errors.SearchFactsError
+import io.dotanuki.norris.networking.NetworkingError
+import io.dotanuki.norris.networking.RemoteServiceIntegrationError.RemoteSystem
 import io.dotanuki.norris.sharedassets.R
 
 data class ErrorStateResources(
