+++ platform/domain/src/test/java/io/dotanuki/norris/domain/FetchCategoriesTests.kt
@@ -1,7 +1,6 @@
 package io.dotanuki.norris.domain
 
 import com.google.common.truth.Truth.assertThat
-import io.dotanuki.norris.domain.errors.NetworkingError
 import io.dotanuki.norris.domain.model.ChuckNorrisFact
 import io.dotanuki.norris.domain.model.RelatedCategory
 import io.dotanuki.norris.domain.services.CategoriesCacheService
@@ -9,6 +8,7 @@ import io.dotanuki.norris.domain.services.RemoteFactsService
 import kotlinx.coroutines.runBlocking
 import org.junit.Before
 import org.junit.Test
+import java.io.IOException
 
 class FetchCategoriesTests {
 
@@ -72,7 +72,7 @@ class FetchCategoriesTests {
     private fun `given that remote service not available`() {
         remoteFacts = object : RemoteFactsService {
             override suspend fun availableCategories(): List<RelatedCategory.Available> {
-                throw NetworkingError.HostUnreachable
+                throw IOException("Timeout")
             }
 
             override suspend fun fetchFacts(searchTerm: String): List<ChuckNorrisFact> = emptyList()
