+++ platform/networking/src/main/java/io/dotanuki/norris/networking/ErrorTransformer.kt
@@ -1,4 +1,4 @@
-package io.dotanuki.norris.domain.errors
+package io.dotanuki.norris.networking
 
 interface ErrorTransformer {
