+++ platform/networking/src/main/java/io/dotanuki/norris/networking/HttpIntegrationErrorTransformer.kt
@@ -1,7 +1,5 @@
 package io.dotanuki.norris.networking
 
-import io.dotanuki.norris.domain.errors.ErrorTransformer
-import io.dotanuki.norris.domain.errors.RemoteServiceIntegrationError
 import retrofit2.HttpException
 
 object HttpIntegrationErrorTransformer : ErrorTransformer {
